pyproject.toml
[tool.poetry]
name = "content-hub"
version = "0.1.0"
description = "collects varying content across several popular websites and displays them all in one place for easy convience and enjoyment"
authors = ["Guido Maldonado <guidojmaldonado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~5.0.1"
django-apscheduler = "~0.6.2"
feedparser = "~6.0.11"
python-dateutil = "~2.8.2"
django-environ = "^0.11.2"
requests = "~2.31.0"
isodate = "~0.6.1"
praw = "~7.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
